+++ CHANGES.md
@@ -8,9 +8,14 @@ Version 1.0.16
 
 To be released.
 
+ -  Fixed a bug where the `fedify init` command had failed to locate package
+    managers on Windows.  [[#210]]
+
  -  The `fedify` command became aware of `FEDIFY_LOG_FILE` environment variable
     to log messages to a file.  If the variable is set, the command logs
     messages to the file specified by the variable.
+
+[#210]: https://github.com/fedify-dev/fedify/issues/210
 
 
 Version 1.0.15